+++ app/controllers/character_log_entries_controller.rb
@@ -136,9 +136,6 @@ class CharacterLogEntriesController < LogEntriesController
                   :downtime_gained, :num_secret_missions,
                   :location_played, :dm_name, :dm_dci_number,
                   :player_dm_id, :notes,
-                  magic_items_attributes: [:id, :name, :rarity,
-                                           :location_found, :table,
-                                           :table_result, :character_id,
-                                           :notes, :_destroy])
+                  magic_items_attributes: magic_item_params)
   end
 end