+++ app/controllers/trade_log_entries_controller.rb
@@ -109,7 +109,6 @@ class TradeLogEntriesController < LogEntriesController
   def log_entries_params
     params.require(:trade_log_entry)
           .permit(:date_played, :downtime_gained, :gp_gained, :traded_magic_item,
-                  :notes, magic_items_attributes: [:id, :name, :rarity, :character_id,
-                                                   :notes, :_destroy])
+                  :notes, magic_items_attributes: magic_item_params)
   end
 end