+++ db/migrate/20170330132511_add_not_included_in_count_to_magic_items.rb
@@ -1,0 +1,5 @@
+class AddNotIncludedInCountToMagicItems < ActiveRecord::Migration
+  def change
+    add_column :magic_items, :not_included_in_count, :boolean, default: false, null: false
+  end
+end